Patch import config when is not valid
